refactored chat server, decoupled room/server/user
**This is currently NOT WORKING still** as a lot of the init logic has
changed, and i need to do yet another pass to appease the lint gods.
It does compile, and the spaghetti mess has been mostly dealt with.

here's some high-level details of the refactor:
* broke logic out of main, moved to `cmd/devzat-server` as build target
* bubble up errors, dont fmt.Print unless necessary
* everything has been decoupled through interfaces declared in
pkg/interfaces
* high-level interfaces for user, server, etc are composed of
un-exported lower-level interfaces in pkg/interfaces
* all commands are now decoupled from the rest of the code. There is an
implicit coupling to the user interface, as the command context is based
of the issuer of the command.
* the `devzat-server` config file is now a json, and it tries to read/write to a
standard place (like `~/.config` or `%APPDATA%`)